+++ CHANGES.md
@@ -3,6 +3,7 @@
  - [BC-BREAK] Drop PSR-11 support.
  - [BC-BREAK] Require PHP 7.2 minimum.
  - [BC-BREAK] Doesn't resolve parameters for factories anymore.
+ - [BC-BREAK] Replace `Registry::withDelegate` with `Registry::setDelegate`.
  - New `Registry::setReflector` to assign a function to cache the creation of `Reflection` objects.
 
 ### 0.13.0